add error catching in controllers

+++ MyLearn/MyLearn/Models/UserCode.cs
@@ -2,6 +2,6 @@
 {
     public class UserCode
     {
-        public float UserTypeCode { get; set; }
+        public int UserTypeCode { get; set; }
     }
 }
